web/auth: split email auth into it’s own component
--- sturdy ---
user_id: cc33c18b-27e5-470d-8589-f9391ca37590
view_id: 2c064d23-2410-4e86-939f-91c56dae2fae
workspace_id: d02f6be6-48c2-4fa1-b852-27055ee4fbc1
