+++ web/src/atoms/auth/SingleOtpInput.vue
