+++ web/src/atoms/auth/index.ts
@@ -1,0 +1,1 @@
+export { default as SingleOtpInput } from './SingleOtpInput.vue'
