+++ web/src/molecules/auth/index.ts
@@ -1,0 +1,1 @@
+export { default as OtpInput } from './OtpInput.vue'
