+++ web/src/organisms/auth/index.ts
@@ -1,0 +1,1 @@
+export { default as EmailAuth } from './Email.vue'
